requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.0
Faker==9.8.3
h11==0.12.0
idna==3.3
loguru==0.5.3
outcome==1.1.0
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
text-unidecode==1.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0